cleanup fuzz closure since we'll use it for merge and cmin as well
